requirements.txt
flask 
sklearn
pymongo
numpy
scipy
geopy
nltk
spacy
pyenchant
BeautifulSoup4
gensim
unidecode
sumy
